+++ CotEditor/Sources/Document Window/File Browser/FileBrowserViewController+QuickLook.swift
@@ -28,20 +28,6 @@ import QuickLookUI
 
 extension FileBrowserViewController {
     
-    override func keyDown(with event: NSEvent) {
-        
-        if event.modifierFlags.isDisjoint(with: .deviceIndependentFlagsMask),
-           event.charactersIgnoringModifiers == " "
-        {
-            // open the Quick Look panel by pressing the Space key
-            self.quickLook(with: event)
-            
-        } else {
-            super.keyDown(with: event)
-        }
-    }
-    
-    
     override func quickLook(with event: NSEvent) {
         
         self.quickLookPreviewItems(nil)